13/9/2017(Try to Decrypt file create to setup File)

+++ ExamSimulator/MainWindow.xaml.cs
@@ -41,7 +41,7 @@ namespace ExamSimulator
             if (Directory.Exists(System.AppDomain.CurrentDomain.BaseDirectory + "\\Examfile\\"))
             {
                 //var allowedExtensions = new[] { ".doc", ".docx" };
-                var allowedExtensions = new[] { ".vcee"};
+                var allowedExtensions = new[] { ".vcee" };
                 var filesss = Directory
                     .GetFiles(System.AppDomain.CurrentDomain.BaseDirectory + "\\Examfile\\")
                     .Where(file => allowedExtensions.Any(file.ToLower().EndsWith))
@@ -81,9 +81,12 @@ namespace ExamSimulator
         private void TriggerClose(object sender, RoutedEventArgs e)
         {
             System.IO.DirectoryInfo di = new DirectoryInfo(System.AppDomain.CurrentDomain.BaseDirectory + "\\ExamReadfile\\");
-            foreach (FileInfo file in di.GetFiles())
+            if (di.GetFiles().Any())
             {
-                file.Delete();
+                foreach (FileInfo file in di.GetFiles())
+                {
+                    file.Delete();
+                }
             }
             Application.Current.Shutdown();
         }
